+++ modules/client/codec/circe/src/main/scala/guara/clientCodec.scala
@@ -1,15 +1,9 @@
-package guara
+package guara.http.client.codec.circe
 
 import io.circe.*
 import io.circe.syntax.*
 import io.circe.parser.{decode as circeDecode}
 import guara.http.client.*
 
-object clientCodecCirce {
-
-  given encoderFromCirce[T](using enc: Encoder[T]): RequestEncoder[T] =
-    (value: T) => Right(value.asJson.noSpaces)
-
-  given decoderFromCirce[T](using dec: Decoder[T]): ResponseDecoder[T] =
-    (text: String) => circeDecode[T](text).left.map(_.getMessage)
-}
+given encoderFromCirce[T](using enc: Encoder[T]): RequestEncoder[T]  = (value: T)     => Right(value.asJson.noSpaces)
+given decoderFromCirce[T](using dec: Decoder[T]): ResponseDecoder[T] = (text: String) => circeDecode[T](text).left.map(_.getMessage)