+++ modules/client/codec/zio/src/main/scala/guara/clientCodec.scala
@@ -1,13 +1,7 @@
-package guara
+package guara.http.client.codec.zio
 
 import zio.json.*
 import guara.http.client.*
 
-object clientCodecZio {
-
-  given encoderFromZioJson[T](using enc: JsonEncoder[T]): RequestEncoder[T] =
-    (value: T) => Right(value.toJson)
-
-  given decoderFromZioJson[T](using dec: JsonDecoder[T]): ResponseDecoder[T] =
-    (text: String) => text.fromJson[T]
-}
+given encoderFromZioJson[T](using enc: JsonEncoder[T]): RequestEncoder[T]  = (value: T)     => Right(value.toJson)
+given decoderFromZioJson[T](using dec: JsonDecoder[T]): ResponseDecoder[T] = (text: String) => text.fromJson[T]